+++ src/reg_cdp1802.cpp
@@ -37,7 +37,7 @@ RegName parseRegName(StrScanner &scan) {
 }
 
 StrBuffer &outRegName(StrBuffer &out, RegName name) {
-    return out.letter('R').uint8(int8_t(name));
+    return out.letter('R').int16(int8_t(name));
 }
 
 }  // namespace reg
